pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "docuextract-ai"
version = "0.1.0"
description = "Local-first document extraction with vision LLMs"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "duckdb>=1.1.0",
    "pymupdf>=1.25.0",
    "pdfplumber>=0.11.0",
    "pdf2image>=1.17.0",
    "ollama>=0.4.0",
    "google-generativeai>=0.8.0",
    "python-multipart>=0.0.18",
    "aiofiles>=24.0.0",
    "pydantic>=2.0.0",
    "pillow>=10.0.0",
]

[tool.setuptools]
packages = ["backend"]

[tool.setuptools.package-data]
"*" = ["*.txt", "*.md"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "httpx>=0.28.0",
]
